[#159587602] Errors wrapping * Rewrited unit tests to use testify * Fixed gometalinter yet anohter time * Removed logrus dependency from default console logger to avoid situations when plugins are using logrus * Separated plugin loader to user correct skip depth

+++ plugins/device/light.go
@@ -22,7 +22,7 @@ type ILight interface {
 
 // LightState returns information about known light.
 type LightState struct {
-	TransitionTime    uint16       `json:"transition_time"`
+	TransitionTime    int          `json:"transition_time"`
 	BrightnessPercent uint8        `json:"brightness"`
 	On                bool         `json:"on"`
 	Color             common.Color `json:"color"`
